feat(address-service): DOMA-4064 Providers detectors now functions in single module
